cmd/filter/requirements.txt
click           >= 6.0
mysql-connector == 2.1.6
pyyaml          >= 3.0
